Make Wsdl2JavaTask @CacheableTask, determine @InputFiles, register tasks in extension
Calling setArgs in a task action makes the whole action have no inputs
during configuration time, which leads to the task always being up to
date. To prevent realization issues (#7), the tasks are defined in the
extension when the container is configured, so realizing them at any point
should just work (as they have all the arguments set right away).