Make usage of binding files compatible with configuration cache
Prior to this commit, the original implementation referenced Project from the task directly.
This resulted in an error during task execution when configuration cache is enabled.

A ProjectLayout is injected into the task and passed along to the arguments provider.
Additionally, the task itself is also passed along which allows the arguments provider class to be static.

Closes gh-98